add root, layout tweaks, couple fixes

+++ src/coreclr/tools/aot/ILCompiler.RyuJit/JitInterface/CorInfoImpl.RyuJit.cs
@@ -2145,7 +2145,7 @@ namespace Internal.JitInterface
                     {
                         pResult->helper = CorInfoHelpFunc.CORINFO_HELP_READYTORUN_THREADSTATIC_BASE;
                         helperId = ReadyToRunHelperId.GetThreadStaticBase;
-                        fieldOffset += (uint)_compilation.NodeFactory.ThreadStaticBaseOffset((MetadataType)field.OwningType);
+                        fieldOffset += _compilation.NodeFactory.ThreadStaticBaseOffset((MetadataType)field.OwningType);
                     }
                     else if (!_compilation.HasLazyStaticConstructor(field.OwningType))
                     {
